requirements.txt
pyside2
gphoto2
numpy
opencv-python
pyserial
RPi.GPIO